packages/dashboard-server/pyproject.toml
[tool.poetry]
name = "dashboard-server"
version = "0.2.0"
description = "FastAPI Boilerplate"
package-mode=false
authors = ["Hide <padocon@naver.com>"]

[tool.poetry.dependencies]
python = "3.11.12"
alembic = "^1.13.1"
uvicorn = "^0.25.0"
fastapi = "^0.109.1"
celery = "^5.3.6"
gunicorn = "^23.0.0"
fastapi-event = "^0.1.3"
pythondi = "^1.2.4"
ujson = "^5.9.0"
aiomysql = "^0.2.0"
click = "^8.1.7"
redis = "^5.0.1"
pydantic-settings = "^2.1.0"
pyjwt = "^2.8.0"
pytest-asyncio = "^0.23.3"
cryptography = "^41.0.7"
coverage = "^7.4.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.25"}
pymysql = "^1.1.0"
dependency-injector = "^4.41.0"
httpx = "^0.26.0"
pymongo = "^4.16.0"
sentry-sdk = {extras = ["fastapi"], version = "^2.51.0"}

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
types-redis = "^4.6.0.20240106"

[tool.coverage.run]
omit=["tests/*", "test_*.py", "migrations/*"]

[tool.isort]
profile="black"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
